Use module structure
All exported class and functions inside index.js are mirrored
into their respective module. These modules are then exported
like they were before.
