+++ src/core/ImageExtend.js
@@ -1,3 +1,4 @@
+import {QuillWatch} from './QuillWatch'
 /**
  * @description 图片功能拓展： 增加上传 拖动 复制
  */
@@ -14,23 +15,12 @@ export class ImageExtend {
         this.config = config
         this.file = ''  // 要上传的图片
         this.imgURL = ''  // 图片地址
-        let quillLoading
-        if (this.config.loading) {
-            quillLoading = document.querySelector('.quillLoading')
-            if (quillLoading === null) {
-                quillLoading = document.createElement('div')
-                quillLoading.classList.add('quill-loading')
-                quillLoading.classList.add('extend-upload-success')
-                document.querySelector('.ql-container').appendChild(quillLoading)
-            }
-        }
-        this.quillLoading = quillLoading
         quill.root.addEventListener('paste', this.pasteHandle.bind(this), false)
         quill.root.addEventListener('drop', this.dropHandle.bind(this), false)
         quill.root.addEventListener('dropover', function (e) {
             e.preventDefault()
         }, false)
-
+        this.cursorIndex = 0
         QuillWatch.on(this.id, this)
     }
 
@@ -39,6 +29,8 @@ export class ImageExtend {
      * @param e
      */
     pasteHandle(e) {
+        // e.preventDefault()
+        QuillWatch.emit(this.quill.id, 0)
         let clipboardData = e.clipboardData
         let i = 0
         let items, item, types
@@ -62,20 +54,16 @@ export class ImageExtend {
                 this.file = item.getAsFile()
                 let self = this
                 // 如果图片限制大小
-                if (self.config.size && self.file.size >= self.config.size * 1024 * 1024 && self.config.loading) {
-                    self.quillLoading.classList.remove('extend-upload-success')
-                    self.quillLoading.classList.add('extend-upload-warning-color')
-                    self.quillLoading.innerHTML = '图片大小超过限制'
-                    setTimeout(function () {
-                        self.quillLoading.classList.remove('extend-upload-warning-color')
-                        self.quillLoading.classList.add('extend-upload-success')
-                    }, 1500)
+                if (self.config.size && self.file.size >= self.config.size * 1024 * 1024) {
+                    if (self.config.sizeError) {
+                        self.config.sizeError()
+                    }
                     return
                 }
                 if (this.config.action) {
-                    this.uploadImg()
+                    // this.uploadImg()
                 } else {
-                    this.toBase64()
+                    // this.toBase64()
                 }
             }
         }
@@ -86,17 +74,14 @@ export class ImageExtend {
      * @param e
      */
     dropHandle(e) {
+        QuillWatch.emit(this.quill.id, 0)
         const self = this
         e.preventDefault()
         // 如果图片限制大小
-        if (self.config.size && self.file.size >= self.config.size * 1024 * 1024 && self.config.loading) {
-            self.quillLoading.classList.add('extend-upload-warning-color')
-            self.quillLoading.classList.remove('extend-upload-success')
-            self.quillLoading.innerHTML = '图片大小超过限制'
-            setTimeout(function () {
-                self.quillLoading.classList.remove('extend-upload-warning-color')
-                self.quillLoading.classList.add('extend-upload-success')
-            }, 1500)
+        if (self.config.size && self.file.size >= self.config.size * 1024 * 1024) {
+            if (self.config.sizeError) {
+                self.config.sizeError()
+            }
             return
         }
         self.file = e.dataTransfer.files[0]; // 获取到第一个上传的文件对象
@@ -145,29 +130,31 @@ export class ImageExtend {
         if (config.change) {
             config.change(xhr, formData)
         }
-        xhr.onload = function (e) {
-            if (self.config.loading) {
-                setTimeout(function () {
-                    quillLoading.classList.add('extend-upload-success')
-                }, 1000)
-            }
-            if (xhr.status === 200) {
-                self.quill.root.innerHTML = self.quill.root.innerHTML.replace('[uploading...]', '')
-                let res = JSON.parse(xhr.responseText)
-                self.imgURL = config.response(res)
-                self.insertImg()
-            } else {
-                self.quill.root.innerHTML = self.quill.root.innerHTML.replace('[uploading...]', '[upload error]')
+        xhr.onreadystatechange = function () {
+            if (xhr.readyState === 4) {
+                if (xhr.status === 200) {
+                    //success
+                    let res = JSON.parse(xhr.responseText)
+                    self.imgURL = config.response(res)
+                    QuillWatch.active.uploadSuccess()
+                    self.insertImg()
+                    if (self.config.success) {
+                        self.config.success()
+                    }
+                } else {
+                    //error
+                    if (self.config.error) {
+                        self.config.error()
+                    }
+                    QuillWatch.active.uploadError()
+                }
             }
         }
         // 开始上传数据
         xhr.upload.onloadstart = function (e) {
-
-            let length = (self.quill.getSelection() || {}).index || self.quill.getLength()
-            self.quill.insertText(length, '[uploading...]', { 'color': 'red'}, true)
-            if (self.config.loading) {
-                quillLoading.classList.remove('extend-upload-success')
-            }
+            QuillWatch.active.uploading()
+            // let length = (self.quill.getSelection() || {}).index || self.quill.getLength()
+            // self.quill.insertText(length, '[uploading...]', { 'color': 'red'}, true)
             if (config.start) {
                 config.start()
             }
@@ -175,21 +162,11 @@ export class ImageExtend {
         // 上传过程
         xhr.upload.onprogress = function (e) {
             let complete = (e.loaded / e.total * 100 | 0) + '%'
-            if (self.config.loading) {
-                quillLoading.innerHTML = '已上传' + complete
-            }
+            QuillWatch.active.progress(complete)
         }
         // 当发生网络异常的时候会触发，如果上传数据的过程还未结束
         xhr.upload.onerror = function (e) {
-            self.quill.root.innerHTML = self.quill.root.innerHTML.replace('[uploading...]', '[upload error]')
-            if (self.config.loading) {
-                self.quillLoading.classList.add('extend-upload-warning-color')
-                quillLoading.innerHTML = '网络异常，失败请重试'
-                setTimeout(function () {
-                    self.quillLoading.classList.remove('extend-upload-warning-color')
-                    quillLoading.classList.add('extend-upload-success')
-                }, 1500)
-            }
+            QuillWatch.active.uploadError()
             if (config.error) {
                 config.error()
             }
@@ -207,11 +184,41 @@ export class ImageExtend {
      * @description 往富文本编辑器插入图片
      */
     insertImg() {
-        const self = this
-        self.quill.blur()
-        let length = (this.quill.getSelection() || {}).index || this.quill.getLength()
-        self.quill.insertEmbed(length, 'image', self.imgURL, 'user')
-        self.quill.setSelection(length + 1)
+        const self = QuillWatch.active
+        self.quill.insertEmbed(QuillWatch.active.cursorIndex, 'image', self.imgURL)
+        self.quill.update()
+        self.quill.setSelection(self.cursorIndex+1);
     }
 
+    /**
+     * @description 显示上传的进度
+     */
+    progress(pro) {
+        pro = '[' + 'uploading' + pro + ']'
+        QuillWatch.active.quill.root.innerHTML
+            = QuillWatch.active.quill.root.innerHTML.replace(/\[uploading.*?\]/, pro)
+    }
+
+    /**
+     * 开始上传
+     */
+    uploading() {
+        let length = (QuillWatch.active.quill.getSelection() || {}).index || QuillWatch.active.quill.getLength()
+        QuillWatch.active.cursorIndex = length
+        QuillWatch.active.quill.insertText(QuillWatch.active.cursorIndex, '[uploading...]', {'color': 'red'}, true)
+    }
+
+    /**
+     * 上传失败
+     */
+    uploadError() {
+        QuillWatch.active.quill.root.innerHTML
+            = QuillWatch.active.quill.root.innerHTML.replace(/\[uploading.*?\]/, '[upload error]')
+    }
+
+    uploadSuccess() {
+        QuillWatch.active.quill.root.innerHTML
+            = QuillWatch.active.quill.root.innerHTML.replace(/\[uploading.*?\]/, '')
+    }
 }
+
