+++ src/core/QuillWatch.js
@@ -5,13 +5,15 @@ import {imgHandler} from './imgHandler'
 export const QuillWatch = {
     watcher: {},  // 登记编辑器信息
     active: null,  // 当前触发的编辑器
-    on: function(imageExtendId, ImageExtend) {  // 登记注册使用了ImageEXtend的编辑器
+    on: function (imageExtendId, ImageExtend) {  // 登记注册使用了ImageEXtend的编辑器
         if (!this.watcher[imageExtendId]) {
             this.watcher[imageExtendId] = ImageExtend
         }
     },
-    emit: function(activeId) {  // 事件发射触发
-        imgHandler()
+    emit: function (activeId, type = 1) {  // 事件发射触发
         this.active = this.watcher[activeId]
+        if (type === 1) {
+            imgHandler()
+        }
     }
 }
