+++ src/core/imgHandler.js
@@ -1,4 +1,4 @@
-import {QuillWatch} from 'QuillWatch'
+import {QuillWatch} from './QuillWatch'
 /**
  * @description 点击图片上传
  */
@@ -15,18 +15,14 @@ export function imgHandler() {
             self.file = fileInput.files[0]
             fileInput.value = ''
             // 如果图片限制大小
-            if (self.config.size && self.file.size >= self.config.size * 1024 * 1024 && self.config.loading) {
-                self.quillLoading.classList.remove('extend-upload-success')
-                self.quillLoading.classList.add('extend-upload-warning-color')
-                self.quillLoading.innerHTML = '图片大小超过限制'
-                setTimeout(function () {
-                    self.quillLoading.classList.remove('extend-upload-warning-color')
-                    self.quillLoading.classList.add('extend-upload-success')
-                }, 1500)
+            if (self.config.size && self.file.size >= self.config.size * 1024 * 1024) {
+                if (self.config.sizeError) {
+                    self.config.sizeError()
+                }
                 return
             }
             if (self.config.action) {
-                self.uploadImg(self.config.change)
+                self.uploadImg()
             } else {
                 self.toBase64()
             }
